+++ packages/calendar/src/components/calendar-types/date-range-calendar/DateRangeCalendar.stories.tsx
@@ -1,4 +1,3 @@
-import { Box } from "@stenajs-webui/core";
 import { Store, withState } from "@dump247/storybook-state";
 import {
   DateRangeCalendar,
@@ -6,9 +5,6 @@ import {
   DateRangeFocusedInput,
   setDayStateValue,
   useDateRangeCalendarState,
-  travelCalendarTheme,
-  dayHighlightSelect,
-  ExtraDayContentProps,
 } from "@stenajs-webui/calendar";
 import { addDays } from "date-fns";
 import * as React from "react";
@@ -66,15 +62,11 @@ export const Standard = withState<State>({
   />
 ));
 
-Standard.storyName = "standard";
-
 Standard.parameters = {
   notes: { markdown },
 };
 
 export const WithStateHook = () => <DateRangeCalendarWithState />;
-
-WithStateHook.storyName = "with state hook";
 
 export const WithTodayHighlighted = withState<State>({
   startDate: undefined,
@@ -91,8 +83,6 @@ export const WithTodayHighlighted = withState<State>({
     setFocusedInput={(focusedInput) => store.set({ focusedInput })}
   />
 ));
-
-WithTodayHighlighted.storyName = "with today highlighted";
 
 export const WithDefaultHighlights = withState<State>({
   startDate: undefined,
@@ -113,8 +103,6 @@ export const WithDefaultHighlights = withState<State>({
   );
 });
 
-WithDefaultHighlights.storyName = "with default highlights";
-
 export const WithMultipleMonths = withState<State>({
   startDate: undefined,
   endDate: undefined,
@@ -130,8 +118,6 @@ export const WithMultipleMonths = withState<State>({
     setFocusedInput={(focusedInput) => store.set({ focusedInput })}
   />
 ));
-
-WithMultipleMonths.storyName = "with multiple months";
 
 export const WithMultipleRows = withState<State>({
   startDate: undefined,
@@ -149,69 +135,3 @@ export const WithMultipleRows = withState<State>({
     setFocusedInput={(focusedInput) => store.set({ focusedInput })}
   />
 ));
-
-WithMultipleRows.storyName = "with multiple rows";
-
-export const Travel = withState<State>({
-  startDate: undefined,
-  endDate: undefined,
-  focusedInput: "startDate",
-})(({ store }: { store: Store<State> }) => {
-  const renderDayContent: React.ComponentType<ExtraDayContentProps<{}>> = ({
-    dayState,
-  }) =>
-    dayHighlightSelect<JSX.Element | null>(
-      dayState,
-      [],
-      ["today", "enabled"],
-      [
-        <Box
-          position={"absolute"}
-          top={"3px"}
-          left={"3px"}
-          style={{
-            backgroundColor: "#ea143d",
-            width: "6px",
-            height: "6px",
-            borderRadius: "50%",
-          }}
-        />,
-        null,
-      ],
-      <Box
-        position={"absolute"}
-        top={0}
-        left={0}
-        bottom={0}
-        right={0}
-        background="#949494"
-        width="70%"
-        height="1px"
-        style={{
-          margin: "auto",
-          transform: "rotate(-45deg)",
-        }}
-      />
-    ) as JSX.Element;
-  return (
-    <DateRangeCalendar
-      highlightToday
-      startDate={store.state.startDate}
-      endDate={store.state.endDate}
-      focusedInput={store.state.focusedInput}
-      setStartDate={(startDate) => store.set({ startDate })}
-      setEndDate={(endDate) => store.set({ endDate })}
-      setFocusedInput={(focusedInput) => store.set({ focusedInput })}
-      defaultHighlights={["disabled"]}
-      theme={travelCalendarTheme}
-      statePerMonth={statePerMonthWithTwoWeeksEnabled}
-      extraDayContent={renderDayContent}
-    />
-  );
-});
-
-Travel.storyName = "travel";
-
-Travel.parameters = {
-  notes: { markdown },
-};